+++ game.cpp
@@ -20,20 +20,20 @@ void ArcanoidGame::showStartWindow(){
 
   QGraphicsTextItem *titleText = new QGraphicsTextItem(QString("Arcanoid"));
   titleText->setFont(QFont("comic sans",50));
-  int txPos = this->width()/2 - titleText->boundingRect().width()/2;
+  double txPos = this->width()/2 - titleText->boundingRect().width()/2;
   int tyPos = 150;
   titleText->setPos(txPos,tyPos);
   gameScene_->addItem(titleText);
 
   Button *playButton = new Button(QString("New game"));
-  int bxPos = this->width()/2 - playButton->boundingRect().width()/2;
+  double bxPos = this->width()/2 - playButton->boundingRect().width()/2;
   int byPos = 275;
   playButton->setPos(bxPos,byPos);
   connect(playButton,SIGNAL(clicked()),this,SLOT(start()));
   gameScene_->addItem(playButton);
 
   Button *quitButton = new Button(QString("Exit"));
-  int qxPos = this->width()/2 - quitButton->boundingRect().width()/2;
+  double qxPos = this->width()/2 - quitButton->boundingRect().width()/2;
   int qyPos = 350;
   quitButton->setPos(qxPos,qyPos);
   connect(quitButton,SIGNAL(clicked()),this,SLOT(close()));
@@ -106,7 +106,7 @@ void ArcanoidGame::start(){
 
 void ArcanoidGame::gameOver(QString textToDisplay){
   Button *playAgain = new Button(QString("Play again"));
-  int bxPos = this->width()/2 - playAgain->boundingRect().width()/2;
+  double bxPos = this->width()/2 - playAgain->boundingRect().width()/2;
   int byPos = 275;
   playAgain->setPos(bxPos, byPos);
   gameScene_->addItem(playAgain);
@@ -121,7 +121,7 @@ void ArcanoidGame::gameOver(QString textToDisplay){
   QGraphicsTextItem *overText = new QGraphicsTextItem(textToDisplay);
   overText->setFont(QFont("comic sans",50));
   overText->setDefaultTextColor(Qt::white);
-  int txPos = this->width()/2 - overText->boundingRect().width()/2;
+  double txPos = this->width()/2 - overText->boundingRect().width()/2;
   int tyPos = 150;
   overText->setPos(txPos,tyPos);
   gameScene_->addItem(overText);
